requirements.txt
twitchAPI==4.2.0
python-dotenv==1.0.1
text-to-speech==1.6.1
playsound==1.3.0
gtts==2.5.1